Массив результатов в отдельный компонент.
